gobject.Signals.connectData should return gulong - closes ticket #97
git-svn-id: http://svn.dsource.org/projects/gtkd/trunk@810 853fb775-85e9-0310-89d4-f5ae88729b16

+++ src/gobject/Signals.d
@@ -118,7 +118,7 @@ public class Signals
 {
 	
 	/** */
-	public static uint connectData(void* instanc, string detailedSignal, GCallback cHandler, Object data, GClosureNotify destroyData, GConnectFlags connectFlags)
+	public static gulong connectData(void* instanc, string detailedSignal, GCallback cHandler, Object data, GClosureNotify destroyData, GConnectFlags connectFlags)
 	{
 		// gulong g_signal_connect_data (gpointer instance,  const gchar *detailed_signal,  GCallback c_handler,  gpointer data,  GClosureNotify destroy_data,  GConnectFlags connect_flags);
 		return g_signal_connect_data(instanc, Str.toStringz(detailedSignal), cHandler, cast(void*)data, destroyData, connectFlags);
